dspark: widen capture-copy size math to size_t (CodeQL overflow)
CodeQL flagged the two dspark capture-copy sites in llama-context.cpp:
'row' was uint32_t, so n_tokens*row and n_outputs*row (feeding the byte
size n_*row*sizeof(float)) were evaluated in 32-bit before widening to
size_t. No overflow at real capture configs (row=n_capture_layers*n_embd
is small), but latent for large captures. Make 'row' size_t so all
downstream size math is 64-bit. No behavior change. libllama builds clean.

+++ src/llama-context.cpp
@@ -1637,7 +1637,7 @@ int llama_context::encode(const llama_batch & batch_inp) {
         if (t_cap) {
             ggml_backend_t backend_c = ggml_backend_sched_get_tensor_backend(sched.get(), t_cap);
             GGML_ASSERT(backend_c != nullptr);
-            const uint32_t row = cparams.n_capture_layers * hparams.n_embd;
+            const size_t row = (size_t) cparams.n_capture_layers * hparams.n_embd;
             GGML_ASSERT(n_tokens*(int64_t) row <= (int64_t) embd_capture.size);
             ggml_backend_tensor_get_async(backend_c, t_cap, embd_capture.data, 0, n_tokens*row*sizeof(float));
         }
@@ -2106,7 +2106,7 @@ int llama_context::decode(const llama_batch & batch_inp) {
             if (t_cap) {
                 ggml_backend_t backend_c = ggml_backend_sched_get_tensor_backend(sched.get(), t_cap);
                 GGML_ASSERT(backend_c != nullptr);
-                const uint32_t row = cparams.n_capture_layers * hparams.n_embd;
+                const size_t row = (size_t) cparams.n_capture_layers * hparams.n_embd;
                 float * embd_capture_out = embd_capture.data + (size_t) n_outputs_prev * row;
                 GGML_ASSERT((n_outputs_prev + n_outputs)*(int64_t) row <= (int64_t) embd_capture.size);
                 ggml_backend_tensor_get_async(backend_c, t_cap, embd_capture_out, 0, n_outputs*row*sizeof(float));
